Properly store lists as JSON
We want to store lists of e.g. links as a JSON object in SQLite.
Currently the code stores them as a comma-separated list, which makes
it harder to query. This changes the logic to store JS objects as
JSON (unless they are Dates, in which case we store the ISO string).