Create full-text search table
This change adds a second SQLite table, `cfgov_fts`, which can be
searched using SQLite FTS5 full-text search extension:

https://www.sqlite.org/fts5.html

This is created as an "external content table", which means that the
main table content is not duplicated into the FTS table, but only
referenced. This should significantly reduce the storage requirements
of creating the FTS table.

A SQLite INSERT trigger is created to ensure that when content is
added to the main table by the crawl, a reference to it gets added
to the FTS table. I haven't created DELETE or UPDATE triggers as we
shouldn't/won't be updating table content once it is written, but that
is something we could do in future if need be:

https://www.sqlite.org/fts5.html#external_content_tables

This change also adopts the use of the sqlite Node package which allows
for async/await interactions with the database.