Add database download link to page footer
This commit adds a link to the page footer to download the SQLite
database file. The link includes the size of the file to be downloaded,
plus a link to instructions to query it in SQLite.

If there is no database file available to the app, the link is hidden.