Viewer: display most recent crawl times
This commit modifies the viewer page HTML to always show when the
last crawl occurred and how long it took.

The header now includes a human-friendly message of when the last
crawl happened ("Last website crawl: 2 months ago"), which links to
the footer which has more detail:

"Last website crawl: April 11, 2022, 4 a.m. - April 11, 2022, 6:18 a.m.
(2 hours, 17 minutes and 59.27 seconds)"

TODO: I've hardcoded a bit of HTML here; when PR 18 comes in it
introduces a separate CSS file which this should be migrated to.