Disable HSTS when crawling with wget
Disable HSTS when crawling websites using wget. If HSTS is enabled,
and wget tries to download a URL starting with http://, it will record
that http:// URL in the WARC even if it gets rewritten to https://
by wget before the request is made. This potentially causes a WARC to
have duplicate URLs. Disabling HSTS ensures that a redirect happens
instead and HTTPS requests are recorded as being HTTPS.