Improve detail page expandables and breadcrumbs
These changes add support for query string parameters to the page
detail view, such that links from the main search results now contain
?query_type&q parameters that tell the detail page what was searched
for. This allows the detail page to open the expandable that is most
appropriate for the search type.

It also modifies how the breadcrumb works on the detail page. Instead
of having custom JS that attempts to highjack clicks on the breadcrumb
link, we change the link to always go to the homepage, and rely on the
user using their back button, as we do in other contexts.