Expand link search to look for urlencoded URLs
This commit modifies the way that link search works so that not only
does it search for the query string provided but also searches for the
urlencoded version of the query string. This supports links that link
to interstitial pages, for example external links on cf.gov:

https://www.consumerfinance.gov/external-site/?ext_url=https%3A%2F%2Fwww.facebook.com%2FCFPB&signature=mLDxq_fH8p2jVLrkPngPqkmwdsl_JEMSCoqZ74gLcCY

Searching for a URL like "https://www.facebook.com" will now match
pages with the above link, even though the link href has been encoded
(https%3A%2F%2f instead of https://).