Limit pages to single domain by default
Currently the crawl result can generate pages that live off of the
www.consumerfinance.gov domain, due to a quirk in the way that wget
works and generates its output.

Specifically, even though we tell wget to only crawl URLs on
www.consumerfinance.gov, if we have redirects that point off the site,
it will crawl those pages. For example:

www.cf.gov/something -> somewhereelse.com/something

This change makes it so these pages are discarded when we convert from
the wget WARC output to the SQLite database.

This behavior can be controlled by a new --multiple-domains /
--no-multiple-domains flag to the warc_to_db management command.