Remove hardcoded wget crawl reject parameters
Currently when we run wget to crawl the website, we use --reject and
--reject-regex to reject certain URLs, specifically files (like *.pdf)
and pages with query string parameters besides ?page (which is the only
query string parameter we want to crawl).

As implemented currently, both of these are written to reject

- a specific list of file extensions: *.css, *.csv, etc.
- a specific list of query string parameters: CatID=, NavCode=, etc.

I originally created both of these lists based on a crawl of the site,
but over time they may become out of date, and it's suboptimal to use
a hardcoded list of rejections when we instead can use wildcards.

For that reason, this commit changes the wget call to use regular
expressions to handle these rejections:

- to reject all filenames:

[^:/]/.*\.[^/]*$

- to reject all query string parameters except for ?page:

([^\?]*(\?([^p]|$)|\?p([^a]|$)|\?pa([^g]|$)|\?pag([^e]|$)|\?page[^=&$]))

Note that by default wget only supports POSIX regular expressions,
which for example doesn't support positive lookahead, which makes these
expressions uglier than they might otherwise have to be.

Additionally, it's not possible to provide multiple --reject-regex
parameters to wget, so these need to be combined into a single one.

I've tested this against the live site and it functions the same as
the previous implementation.