Server configuration and deployment improvements
This commit adds a new Fabric command, `fab configure`, which can be
used to configure a RHEL7 server to run the crawler application. This
command does the following:

1. Install Node
2. Install Yarn
3. Build a modern version of SQLite
4. Install Python 3

This commit also enhances the existing `fab deploy` command to add
Gunicorn logging, plus logrotate to compress and rotate the log files
daily. We will keep the last 90 days of logs.