Fix: Skip pages with missing title tags
Currently if for some reason a page that is otherwise valid (returned
an HTTP 200 response and is of type text/html) is missing a <title> tag,
the warc_to_db command will fail because titles are not optional.

It might be possible to add support for optional titles, but this would
require some changes to the UI (to know how to display links to the
page, for example).

This change skips any pages that show up in the WARC like this.