Add support for RHEL8
This change adds support for running this application on RHEL8.

The fabfile.py configuration and deployment scripts are simplified;
RHEL8 includes Python 3.6 so we no longer need to install it, nor a
different version of SQLite. We can also use Node 18 instead of Node 16.

The Python application code also needs to be modified to use a more
secure version of hashlib.md5 to comply with FIPS mode. See
http://blog.serindu.com/2019/11/12/django-in-fips-mode/ for background
and the implemented workaround.