Update deploy fabfile for RHEL8
This command updates the repo fabfile.py for deploys to RHEL8, from the
previous RHEL7 configuration. The only change has to do with the
location of the crawl database file itself, moving it from /var/tmp
to the deploy location in /opt/cfpb/website-indexer. This change is
necesssary due to the different volume configuration in our RHEL8 AMI.