Implement frontend build using CFPB Design System 2.0.1 (#96)
This PR adds an esbuild-based frontend build to this project using the CFPB Design System release 2.0.1. The previous frontend "build" downloaded compiled CSS from consumerfinance.gov instead of building CSS itself. This broke recently when the cf.gov codebase was modified and the HTML/CSS in this project wasn't updated. The various Django templates in this project have been updated to work properly with the DS 2.0.1.