Deprecate explicit RHEL8 FIPS mode support
PR 89 migrated the EC2 deployment configuration for this repository to
RHEL8. Those changes included some explicit monkeypatching of MD5 to
allow for running on RHEL8 with FIPS mode enabled.

Going forward, support for EC2 deployment will be maintained in the ec2
branch [1] until such time as we are fully containerized.

For that reason, we no longer need to maintain this code in the main
branch, and it can be removed.

[0] https://github.com/cfpb/website-indexer/pull/89
[1] https://github.com/cfpb/website-indexer/tree/ec2