Allow configuration from file-based env variables
As an alternative to configuration via environment variable, this
commit allows for configuration from files instead. This avoids the
need to populate a container environment with secret values directly.

The PATCH_ENVIRON_PATH can be pointed to a directory containing files
named for the variables to set, containing the desired values for those
variables.

The README has been updated to document this feature.