Better cleanup of incomplete crawls
If a crawl job is interrupted, it won't record a final status of
`FAILED` or `FINISHED`, but will be left in `STARTED`. The current
invocation of "manage.py manage_crawls clean" won't clean up old
`STARTED` crawls. This commit modifies the logic of that command to do
so, while still retaining the ability to keep 1 completed `FINISHED`
and `FAILED` crawl, respectively. The logic won't delete `STARTED`
crawls that are the most recent.

Unit tests have been updated to cover this change.