Restore DEBUG=True by default
Partially reverts PR 131. We want to keep DEBUG=True by default when
running the webserver (for easy static files and nice debug errors),
but we still want the ability to disable DEBUG mode when crawling.

Add the ability to disable DEBUG via the DJANGO_DEBUG environment
variable. Our nightly crawl is configured in an internal crawler-deploy
repo that contains the helm chart, so a related change must be made
there as well.